Move dense_vector and sparse_vector to module (#43280)
