--- modules/mapper-extras/src/main/resources/META-INF/services/org.elasticsearch.painless.spi.PainlessExtension
@@ -1,1 +1,0 @@
-org.elasticsearch.index.query.DocValuesWhitelistExtension
