+++ x-pack/plugin/src/test/resources/rest-api-spec/test/vectors/10_dense_vector_basic.yml
