+++ x-pack/plugin/vectors/src/main/resources/META-INF/services/org.elasticsearch.painless.spi.PainlessExtension
@@ -1,0 +1,1 @@
+org.elasticsearch.xpack.vectors.query.DocValuesWhitelistExtension
